requirements.txt
Django==4.2.1
djangorestframework==3.14.0
python-telegram-bot==20.3
requests==2.31.0